Split bp mods
