+++ crates/block-producer/src/server/mod.rs
@@ -1,10 +1,6 @@
-use std::collections::HashMap;
-use std::net::SocketAddr;
-use std::num::NonZeroUsize;
 use std::sync::Arc;
-use std::time::Duration;
 
-use anyhow::{Context, Result};
+use anyhow::Context;
 use futures::StreamExt;
 use miden_node_proto::domain::mempool::MempoolEvent;
 use miden_node_proto::generated::block_producer::api_server;
@@ -23,20 +19,22 @@ use tokio::sync::{Mutex, RwLock};
 use tokio_stream::wrappers::{ReceiverStream, TcpListenerStream};
 use tonic::Status;
 use tower_http::trace::TraceLayer;
-use tracing::{debug, error, info, instrument};
-use url::Url;
+use tracing::{debug, instrument};
 
-use crate::batch_builder::BatchBuilder;
-use crate::block_builder::BlockBuilder;
 use crate::domain::transaction::AuthenticatedTransaction;
-use crate::errors::{BlockProducerError, MempoolSubmissionError, StoreError};
-use crate::mempool::{BatchBudget, BlockBudget, Mempool, MempoolConfig, SharedMempool};
+use crate::errors::MempoolSubmissionError;
+use crate::mempool::SharedMempool;
 use crate::store::StoreClient;
-use crate::validator::BlockProducerValidatorClient;
-use crate::{CACHED_MEMPOOL_STATS_UPDATE_INTERVAL, COMPONENT, SERVER_NUM_BATCH_BUILDERS};
+use crate::{CACHED_MEMPOOL_STATS_UPDATE_INTERVAL, COMPONENT};
+
+pub mod embedded;
+pub mod rpc;
 
 #[cfg(test)]
 mod tests;
+
+pub use embedded::EmbeddedBlockProducer;
+pub use rpc::BlockProducer;
 
 // BLOCK PRODUCER HANDLE
 // ================================================================================================
@@ -51,13 +49,13 @@ mod tests;
 /// block builders equal footing with all incoming transactions combined.
 #[derive(Clone)]
 pub struct BlockProducerHandle {
-    mempool: Arc<Mutex<SharedMempool>>,
-    store: StoreClient,
-    cached_mempool_stats: Arc<RwLock<MempoolStats>>,
+    pub(super) mempool: Arc<Mutex<SharedMempool>>,
+    pub(super) store: StoreClient,
+    pub(super) cached_mempool_stats: Arc<RwLock<MempoolStats>>,
 }
 
 impl BlockProducerHandle {
-    fn new(mempool: SharedMempool, store: StoreClient) -> Self {
+    pub(super) fn new(mempool: SharedMempool, store: StoreClient) -> Self {
         Self {
             mempool: Arc::new(Mutex::new(mempool)),
             store,
@@ -68,7 +66,7 @@ impl BlockProducerHandle {
     /// Starts a background task that periodically updates the cached mempool statistics.
     ///
     /// This prevents the need to lock the mempool for each status request.
-    async fn spawn_stats_updater(&self) {
+    pub(super) async fn spawn_stats_updater(&self) {
         let cached_mempool_stats = Arc::clone(&self.cached_mempool_stats);
         let mempool = self.mempool.lock().await.clone();
 
@@ -207,301 +205,8 @@ impl BlockProducerHandle {
         }))
     }
 
-    async fn subscribe(&self) -> ReceiverStream<MempoolEvent> {
+    pub(super) async fn subscribe(&self) -> ReceiverStream<MempoolEvent> {
         ReceiverStream::new(self.mempool.lock().await.lock().await.subscribe())
-    }
-}
-
-// EMBEDDED BLOCK PRODUCER
-// ================================================================================================
-
-/// Block producer variant that uses an in-process store instead of a remote gRPC store.
-pub struct EmbeddedBlockProducer {
-    /// The address of the block producer component.
-    pub block_producer_address: SocketAddr,
-    /// The in-process store state.
-    pub state: Arc<miden_node_store::State>,
-    /// The address of the validator component.
-    pub validator_url: Url,
-    /// The address of the batch prover component.
-    pub batch_prover_url: Option<Url>,
-    /// The interval at which to produce batches.
-    pub batch_interval: Duration,
-    /// The interval at which to produce blocks.
-    pub block_interval: Duration,
-    /// The maximum number of transactions per batch.
-    pub max_txs_per_batch: usize,
-    /// The maximum number of batches per block.
-    pub max_batches_per_block: usize,
-    /// Server-side gRPC options.
-    pub grpc_options: GrpcOptionsInternal,
-    /// The maximum number of inflight transactions allowed in the mempool at once.
-    pub mempool_tx_capacity: NonZeroUsize,
-}
-
-impl EmbeddedBlockProducer {
-    /// Initialises the block producer internals and returns an in-process [`BlockProducerHandle`]
-    /// together with a future that runs the gRPC server, batch builder, and block builder.
-    ///
-    /// Use this when the caller needs to submit transactions directly to the mempool without a
-    /// gRPC round-trip (e.g. the embedded sequencer's RPC).
-    pub async fn start(
-        self,
-    ) -> anyhow::Result<(BlockProducerHandle, impl Future<Output = anyhow::Result<()>> + Send)>
-    {
-        info!(target: COMPONENT, endpoint=?self.block_producer_address, "Initializing embedded server");
-        let store = StoreClient::new_local(self.state.clone());
-        let validator = BlockProducerValidatorClient::new(self.validator_url.clone());
-
-        let chain_tip = self.state.chain_tip(miden_node_store::Finality::Committed).await;
-
-        let block_builder = BlockBuilder::new(store.clone(), validator, self.block_interval);
-        let batch_builder = BatchBuilder::new(
-            store.clone(),
-            SERVER_NUM_BATCH_BUILDERS,
-            self.batch_prover_url,
-            self.batch_interval,
-        );
-        let mempool = Mempool::shared(
-            chain_tip,
-            MempoolConfig {
-                batch_budget: BatchBudget {
-                    transactions: self.max_txs_per_batch,
-                    ..BatchBudget::default()
-                },
-                block_budget: BlockBudget { batches: self.max_batches_per_block },
-                tx_capacity: self.mempool_tx_capacity,
-                ..Default::default()
-            },
-        );
-
-        let handle = BlockProducerHandle::new(mempool.clone(), store);
-
-        let grpc_options = self.grpc_options;
-        let block_producer_address = self.block_producer_address;
-
-        let serve = {
-            let handle = handle.clone();
-            async move {
-                let listener = TcpListener::bind(block_producer_address)
-                    .await
-                    .context("failed to bind to block producer address")?;
-
-                info!(target: COMPONENT, "Embedded server initialized");
-
-                let mut tasks = tokio::task::JoinSet::new();
-
-                let rpc_id = tasks
-                    .spawn(async move {
-                        BlockProducerRpcServer::new(handle).serve(listener, grpc_options).await
-                    })
-                    .id();
-
-                let batch_builder_id = tasks
-                    .spawn({
-                        let mempool = mempool.clone();
-                        async {
-                            batch_builder.run(mempool).await;
-                            Ok(())
-                        }
-                    })
-                    .id();
-
-                let block_builder_id = tasks.spawn(async { block_builder.run(mempool).await }).id();
-
-                let task_ids = HashMap::from([
-                    (batch_builder_id, "batch-builder"),
-                    (block_builder_id, "block-builder"),
-                    (rpc_id, "rpc"),
-                ]);
-
-                let task_result = tasks.join_next_with_id().await.unwrap();
-
-                let task_id = match &task_result {
-                    Ok((id, _)) => *id,
-                    Err(err) => err.id(),
-                };
-                let task = task_ids.get(&task_id).unwrap_or(&"unknown");
-
-                task_result
-                    .map_err(|source| BlockProducerError::JoinError { task, source })
-                    .map(|(_, result)| match result {
-                        Ok(_) => Err(BlockProducerError::UnexpectedTaskCompletion { task }),
-                        Err(source) => Err(BlockProducerError::TaskError { task, source }),
-                    })
-                    .and_then(|x| x)?
-            }
-        };
-
-        Ok((handle, serve))
-    }
-
-    /// Serves the block-producer RPC API, the batch-builder and the block-builder.
-    ///
-    /// Executes in place (i.e. not spawned) and will run indefinitely until a fatal error is
-    /// encountered.
-    pub async fn serve(self) -> anyhow::Result<()> {
-        let (_handle, serve) = self.start().await?;
-        serve.await
-    }
-}
-
-// BLOCK PRODUCER
-// ================================================================================================
-
-/// The block producer server.
-///
-/// Specifies how to connect to the store, batch prover, and block prover components.
-/// The connection to the store is established at startup and retried with exponential backoff
-/// until the store becomes available. Once the connection is established, the block producer
-/// will start serving requests.
-pub struct BlockProducer {
-    /// The address of the block producer component.
-    pub block_producer_address: SocketAddr,
-    /// The address of the store component.
-    pub store_url: Url,
-    /// The address of the validator component.
-    pub validator_url: Url,
-    /// The address of the batch prover component.
-    pub batch_prover_url: Option<Url>,
-    /// The interval at which to produce batches.
-    pub batch_interval: Duration,
-    /// The interval at which to produce blocks.
-    pub block_interval: Duration,
-    /// The maximum number of transactions per batch.
-    pub max_txs_per_batch: usize,
-    /// The maximum number of batches per block.
-    pub max_batches_per_block: usize,
-    /// Server-side gRPC options.
-    pub grpc_options: GrpcOptionsInternal,
-
-    /// The maximum number of inflight transactions allowed in the mempool at once.
-    pub mempool_tx_capacity: NonZeroUsize,
-}
-
-impl BlockProducer {
-    /// Serves the block-producer RPC API, the batch-builder and the block-builder.
-    ///
-    /// Executes in place (i.e. not spawned) and will run indefinitely until a fatal error is
-    /// encountered.
-    pub async fn serve(self) -> anyhow::Result<()> {
-        info!(target: COMPONENT, endpoint=?self.block_producer_address, store=%self.store_url, "Initializing server");
-        let store = StoreClient::new(self.store_url.clone());
-        let validator = BlockProducerValidatorClient::new(self.validator_url.clone());
-
-        // Retry fetching the chain tip from the store until it succeeds.
-        let mut retries_counter = 0;
-        let chain_tip = loop {
-            match store.latest_header().await {
-                Err(StoreError::GrpcClientError(err)) => {
-                    // exponential backoff with base 500ms and max 30s
-                    let backoff = Duration::from_millis(500)
-                        .saturating_mul(1 << retries_counter)
-                        .min(Duration::from_secs(30));
-
-                    error!(
-                        store = %self.store_url,
-                        ?backoff,
-                        %retries_counter,
-                        %err,
-                        "store connection failed while fetching chain tip, retrying"
-                    );
-
-                    retries_counter += 1;
-                    tokio::time::sleep(backoff).await;
-                },
-                Ok(header) => break header.block_num(),
-                Err(e) => {
-                    error!(target: COMPONENT, %e, "failed to fetch chain tip from store");
-                    return Err(e.into());
-                },
-            }
-        };
-
-        let listener = TcpListener::bind(self.block_producer_address)
-            .await
-            .context("failed to bind to block producer address")?;
-
-        info!(target: COMPONENT, "Server initialized");
-
-        let block_builder = BlockBuilder::new(store.clone(), validator, self.block_interval);
-        let batch_builder = BatchBuilder::new(
-            store.clone(),
-            SERVER_NUM_BATCH_BUILDERS,
-            self.batch_prover_url,
-            self.batch_interval,
-        );
-        let mempool = Mempool::shared(
-            chain_tip,
-            MempoolConfig {
-                batch_budget: BatchBudget {
-                    transactions: self.max_txs_per_batch,
-                    ..BatchBudget::default()
-                },
-                block_budget: BlockBudget { batches: self.max_batches_per_block },
-                tx_capacity: self.mempool_tx_capacity,
-                ..Default::default()
-            },
-        );
-
-        let handle = BlockProducerHandle::new(mempool.clone(), store);
-
-        // Spawn rpc server and batch and block provers.
-        //
-        // These communicate indirectly via a shared mempool.
-        //
-        // These should run forever, so we combine them into a joinset so that if
-        // any complete or fail, we can shutdown the rest (somewhat) gracefully.
-        let mut tasks = tokio::task::JoinSet::new();
-
-        let rpc_id = tasks
-            .spawn(async move {
-                BlockProducerRpcServer::new(handle).serve(listener, self.grpc_options).await
-            })
-            .id();
-
-        let batch_builder_id = tasks
-            .spawn({
-                let mempool = mempool.clone();
-                async {
-                    batch_builder.run(mempool).await;
-                    Ok(())
-                }
-            })
-            .id();
-        let block_builder_id = tasks
-            .spawn({
-                let mempool = mempool.clone();
-                async { block_builder.run(mempool).await }
-            })
-            .id();
-
-        let task_ids = HashMap::from([
-            (batch_builder_id, "batch-builder"),
-            (block_builder_id, "block-builder"),
-            (rpc_id, "rpc"),
-        ]);
-
-        // Wait for any task to end. They should run indefinitely, so this is an unexpected result.
-        //
-        // SAFETY: The JoinSet is definitely not empty.
-        let task_result = tasks.join_next_with_id().await.unwrap();
-
-        let task_id = match &task_result {
-            Ok((id, _)) => *id,
-            Err(err) => err.id(),
-        };
-        let task = task_ids.get(&task_id).unwrap_or(&"unknown");
-
-        // We could abort the other tasks here, but not much point as we're probably crashing the
-        // node.
-        task_result
-            .map_err(|source| BlockProducerError::JoinError { task, source })
-            .map(|(_, result)| match result {
-                Ok(_) => Err(BlockProducerError::UnexpectedTaskCompletion { task }),
-                Err(source) => Err(BlockProducerError::TaskError { task, source }),
-            })
-            .and_then(|x| x)?
     }
 }
 
@@ -511,16 +216,16 @@ impl BlockProducer {
 /// Serves the block producer's RPC [api](api_server::Api).
 ///
 /// A thin gRPC adapter over [`BlockProducerHandle`] — all business logic lives on the handle.
-struct BlockProducerRpcServer {
+pub(super) struct BlockProducerRpcServer {
     handle: BlockProducerHandle,
 }
 
 impl BlockProducerRpcServer {
-    fn new(handle: BlockProducerHandle) -> Self {
+    pub(super) fn new(handle: BlockProducerHandle) -> Self {
         Self { handle }
     }
 
-    async fn serve(
+    pub(super) async fn serve(
         self,
         listener: TcpListener,
         grpc_options: GrpcOptionsInternal,
@@ -583,7 +288,7 @@ impl api_server::Api for BlockProducerRpcServer {
 // MEMPOOL SUBSCRIPTION
 // ================================================================================================
 
-struct MempoolEventSubscription {
+pub(super) struct MempoolEventSubscription {
     inner: ReceiverStream<MempoolEvent>,
 }
 
@@ -605,7 +310,7 @@ impl tokio_stream::Stream for MempoolEventSubscription {
 
 /// Mempool statistics that are updated periodically to avoid locking the mempool.
 #[derive(Clone, Copy, Default)]
-struct MempoolStats {
+pub(super) struct MempoolStats {
     /// The mempool's current view of the chain tip height.
     chain_tip: BlockNumber,
     /// Number of transactions currently in the mempool waiting to be batched.
